refactor: Divided index.ts content in different files with a better and intuitive structure.
