+++ packages/core/src/bindings/setupBindings.ts
@@ -1,0 +1,28 @@
+import type { BindingsCollection } from "./types";
+import { setupValueBindings, renderValueBindings } from "./bindValue";
+import { setupVisibleBindings, renderVisibleBindings } from "./bindVisible";
+import { setupClassBindings, renderClassBindings } from "./bindClass";
+import { setupStyleBindings, renderStyleBindings } from "./bindStyle";
+import { setupClickBindings } from "./bindClick";
+
+export function setupBindings(root: HTMLElement, viewModel: any): () => void {
+  const bindings: BindingsCollection = {
+    valueBindings: setupValueBindings(root, viewModel),
+    visibleBindings: setupVisibleBindings(root, viewModel),
+    classBindings: setupClassBindings(root, viewModel),
+    styleBindings: setupStyleBindings(root, viewModel),
+  };
+
+  setupClickBindings(root, viewModel);
+
+  const render = () => {
+    renderValueBindings(bindings.valueBindings, viewModel);
+    renderVisibleBindings(bindings.visibleBindings, viewModel);
+    renderClassBindings(bindings.classBindings, viewModel);
+    renderStyleBindings(bindings.styleBindings, viewModel);
+  };
+
+  render();
+  return render;
+}
+
