+++ packages/core/src/bootstrap/mountTemplate.ts
@@ -1,0 +1,10 @@
+import { bootstrap } from "./bootstrap";
+
+export function mountTemplate(
+  container: HTMLElement,
+  templateHtml: string,
+): void {
+  container.innerHTML = templateHtml;
+  bootstrap({ root: container });
+}
+
