+++ packages/core/src/types.ts
@@ -1,0 +1,9 @@
+export type ViewModelConstructor<T = unknown> = {
+  new (): T;
+};
+
+export type PelelaOptions = {
+  document?: Document;
+  root?: ParentNode;
+};
+
