+++ packages/core/src/validation/assertViewModelProperty.ts
@@ -1,0 +1,16 @@
+export function assertViewModelProperty(
+  viewModel: any,
+  propertyName: string,
+  kind: string,
+  element: Element,
+): void {
+  if (!(propertyName in viewModel)) {
+    const snippet = element.outerHTML.replace(/\s+/g, " ").trim().slice(0, 100);
+
+    throw new Error(
+      `[pelela] Unknown property "${propertyName}" used in ${kind} on: ${snippet}. ` +
+      `Make sure your view model "${viewModel.constructor.name}" defines it.`,
+    );
+  }
+}
+
